asm: fix relocations and dangling bytes
The 6502 assembler took a different approach to relocations, wherein
jump values were stored in the subsequent byte. In this device, jump
values are stored in the current word.

Additionally, jumps only have a 1-word offset, not 2-bytes.

This necessitates several changes, which are presented here. This
includes not eating the subsequent byte, and only subtracting 1 from the
current PC rather than 2.

Signed-off-by: Sean Cross <sean@xobs.io>